move laymanTerms
